Remove duplicate negative coordinate test
